agregando validaciones

+++ consultas/consultas.js
@@ -2,19 +2,35 @@ import pool from "../config/db.js";
 import { manejoErrores } from '../erroresBBDD/errores.js';
 const tabla = 'canciones';
 
-
+let mensaje;
 // Agregar una nueva cancion
 const nuevaCancion = async (titulo, artista, tono) => {
     try {
+        if (!titulo || !artista || !tono) {
+            mensaje = "Debe ingresar todos los campos: titulo, artista y tono";
+            console.log(mensaje);
+            return mensaje;
+        } else {
+            const consulta = {
+                text: `SELECT * FROM ${tabla} WHERE titulo = $1 AND artista = $2 AND tono = $3`,
+                values: [titulo, artista, tono]
+            }
+
+            const verificar = await pool.query(consulta);
+            if (verificar.rows != 0) {
+                mensaje = "La canción ya existe en la base de datos"
+                console.log(mensaje)
+                return mensaje;
+            }
+        }
         const consulta = {
             text: ` INSERT INTO ${tabla} (titulo, artista, tono) values ($1, $2, $3) RETURNING *`,
             values: [titulo, artista, tono],
         };
         const resultado = await pool.query(consulta);
-
         if (resultado.rows == 0) {
-            console.log("No se pudo agregar la cancion");
-            const mensaje = "No se pudo agregar la cancion";
+            mensaje = "No se pudo agregar la cancion";
+            console.log(mensaje);
             return mensaje;
         } else {
             console.log(`La cancion con titulo: ${titulo} agregada correctamente. ${JSON.stringify(resultado.rows[0])}`);
@@ -34,8 +50,8 @@ const verCanciones = async () => {
         const resultado = await pool.query(consulta);
 
         if (resultado.rows == 0) {
-            console.log(`No se encontraron canciones`);
-            const mensaje = "No se encontraron canciones registradas";
+            mensaje = "No se encontraron canciones registradas";
+            console.log(mensaje);
             return mensaje;
         } else {
             console.log(`El registro actual de canciones es ${JSON.stringify(resultado.rows)}`);
@@ -49,6 +65,11 @@ const verCanciones = async () => {
 // Actualizar una cancion
 const editarCancion = async (id, titulo, artista, tono) => {
     try {
+        if (!id || !titulo || !artista || !tono) {
+            mensaje = "Debe ingresar todos los campos id, titulo, artista y tono";
+            console.log(mensaje);
+            return mensaje;
+        }
         const consulta = {
             text: `UPDATE ${tabla} SET titulo = $1, artista = $2, tono = $3 WHERE id = $4 RETURNING *`,
             values: [titulo, artista, tono, id]
@@ -56,8 +77,8 @@ const editarCancion = async (id, titulo, artista, tono) => {
         const resultado = await pool.query(consulta);
 
         if (resultado.rows == 0) {
-            console.log("No se pudo actualizar la cancion, ya que no existe");
-            const mensaje = "No se pudo actualizar la cancion, ya que no existe"
+            mensaje = "No se pudo actualizar la cancion, ya que no existe"
+            console.log(mensaje);
             return mensaje;
         } else {
             console.log(`Cancion con titulo: ${titulo} actualizada correctamente.`);
@@ -70,6 +91,11 @@ const editarCancion = async (id, titulo, artista, tono) => {
 
 const eliminarCancion = async (id) => {
     try {
+        if (!id) {
+            mensaje = "Debe ingresar el campo id";
+            console.log(mensaje);
+            return mensaje;
+        }
         const consulta = {
             text: `DELETE FROM ${tabla} WHERE id = $1 RETURNING *`,
             values: [id],
@@ -77,8 +103,8 @@ const eliminarCancion = async (id) => {
         const resultado = await pool.query(consulta);
 
         if (resultado.rows == 0) {
-            console.log("No se pudo eliminar la cancion, ya que no existe");
-            const mensaje = "No se pudo eliminar la cancion, ya que no existe"
+            mensaje = "No se pudo eliminar la cancion, ya que no existe"
+            console.log(mensaje);
             return mensaje;
         } else {
             console.log(`Cancion con id: ${id} eliminada correctamente.`);
